Move authentication clases into auth package
