+++ gnd/src/main/java/com/google/android/gnd/MainActivity.java
@@ -30,9 +30,9 @@ import androidx.navigation.fragment.NavHostFragment;
 import com.google.android.gnd.databinding.MainActBinding;
 import com.google.android.gnd.repository.UserRepository;
 import com.google.android.gnd.system.ActivityStreams;
-import com.google.android.gnd.system.AuthenticationManager;
+import com.google.android.gnd.system.auth.AuthenticationManager;
 import com.google.android.gnd.system.SettingsManager;
-import com.google.android.gnd.system.SignInState;
+import com.google.android.gnd.system.auth.SignInState;
 import com.google.android.gnd.ui.common.BackPressListener;
 import com.google.android.gnd.ui.common.EphemeralPopups;
 import com.google.android.gnd.ui.common.Navigator;
